ref: Renames CSVStatementLoader to StatementReader
